fix: reverse component logic
